Generate socials redirects from data file
